Second Commit
A whole truck load of changes. I started from scratch because I forgot
exactly where I was on the old version, so here's the new one. I used _S
as a base template, and included simplegrid.css to help with the layout.
Then I began toiling away at the styles.

The next couple commits will be to include:
-Additional style
-media-queries for a more responsive layout
-Bug fixes
-styling the other page files